+++ Sources/KeyboardKit/Autocomplete/AutocompleteProvider.swift
@@ -15,6 +15,18 @@ import Foundation
 /// an autocomplete operation that returns suggestions for a
 /// text in the current ``locale``.
 ///
+/// Words can be learned using ``learnWord(_:)`` and ignored
+/// using ``ignoreWord(_:)``. Learned words are persisted in
+/// the system, and will be considered as new words from the
+/// point they are learned, while ignored words will just be
+/// ignored by the autocomplete provider.
+///
+/// Not all autocomplete providers support ignoring/learning
+/// words. Check ``canLearnWords`` and ``canIgnoreWords`` to
+/// check whether or not a provider supports it. Also, while
+/// a provider may support ignoring and learning, it may not
+/// be able to provide ``ignoredWords`` and ``learnedWords``.
+///
 /// KeyboardKit does not have a standard provider, as it has
 /// for other services. Instead, a disabled provider will be
 /// used until you register a custom provider, or register a
@@ -63,4 +75,32 @@ public protocol AutocompleteProvider: AnyObject {
 
     /// Make the provider unlearn a certain word.
     func unlearnWord(_ word: String)
+}
+
+public extension AutocompleteProvider {
+    
+    /// Make the provider ignore a certain suggestion.
+    func ignore(_ suggestion: Autocomplete.Suggestion) {
+        ignoreWord(suggestion.text)
+    }
+    
+    /// Make the provider ignore a collection of words.
+    func ignoreWords(_ words: [String]) {
+        words.forEach(ignoreWord)
+    }
+    
+    /// Make the provider learn a certain suggestion.
+    func learn(_ suggestion: Autocomplete.Suggestion) {
+        learnWord(suggestion.text)
+    }
+    
+    /// Remove a certain ignored suggested from the provider.
+    func removeIgnoredSuggestion(_ suggestion: Autocomplete.Suggestion) {
+        removeIgnoredWord(suggestion.text)
+    }
+    
+    /// Make the provider unlearn a certain suggestion.
+    func unlearn(_ suggestion: Autocomplete.Suggestion) {
+        unlearnWord(suggestion.text)
+    }
 }